refactor(codegen): rename opcode module
